release/v3/requirement.txt
Flask==3.0.3
Flask_Cors==5.0.0
langchain==0.3.1
langchain_community==0.3.1
langchain_core==0.3.7
langchain_openai==0.2.1
langchain_pinecone==0.2.0
pinecone==4.0.0
python-dotenv==1.0.1
Werkzeug==3.0.4
langchain_google_genai==2.0.0
pypdf
unstructured
python-docx